the clean up from intellij is applied

+++ src/graphtea/extensions/algs4/StdAudio.java
@@ -151,7 +151,7 @@ public final class StdAudio {
         int n = data.length;
         double[] d = new double[n/2];
         for (int i = 0; i < n/2; i++) {
-            d[i] = ((short) (((data[2*i+1] & 0xFF) << 8) + (data[2*i] & 0xFF))) / ((double) MAX_16_BIT);
+            d[i] = ((short) (((data[2*i+1] & 0xFF) << 8) + (data[2*i] & 0xFF))) / MAX_16_BIT;
         }
         return d;
     }
