+++ src/graphtea/extensions/reports/zagreb/SDD.java
@@ -198,7 +198,7 @@ public class SDD implements GraphReportExtension {
         double alpha = (1 / m) * (Math.floor(m)) * (1 - ((1 / m) * (Math.floor(m / 2))));
         double alpha1 = (m) * (Math.floor(m / 2)) * (1 - ((1 / m) * (Math.floor(m / 2))));
 
-        int diameter = (int) new Diameter().calculate(g);
+        int diameter = new Diameter().calculate(g);
         WienerIndex wi = new WienerIndex();
         double Avg = (n * (n - 1) / 2);
 
@@ -232,7 +232,7 @@ public class SDD implements GraphReportExtension {
         //  v.add(chi-(n/2));
 
         v.add(new NumOfIndSets().calculate(g));
-        v.add((Integer) new NumOfIndSets().calculate(Utils.createLineGraph(g)));
+        v.add(new NumOfIndSets().calculate(Utils.createLineGraph(g)));
         //   v.add(ISI);
         //  v.add((H*M12)/(2*m));
         //  v.add( ((H*M12)/(2*m)) + ((alpha*m*((maxEdge+2)-(minEdge+2))*((maxDeg*maxDeg)-(minDeg*minDeg))) /((maxEdge+2)*(minEdge+2))) );
